backend/requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
psycopg2-binary
alembic

python-dotenv

pydantic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]

redis

httpx

pytest